+++ .eslintrc.json
@@ -1,3 +1,21 @@
 {
-  "extends": ["next/core-web-vitals", "plugin:@next/next/recommended"]
+    "parser": "@typescript-eslint/parser",
+    "extends": ["next/core-web-vitals", "plugin:@next/next/recommended", "plugin:@typescript-eslint/recommended"],
+    "plugins": ["unused-imports"],
+    "rules": {
+    "no-unused-vars": "off", // or "@typescript-eslint/no-unused-vars": "off",
+    "unused-imports/no-unused-imports": "error",
+    "unused-imports/no-unused-vars": [
+        "warn",
+        {
+            "vars": "all",
+            "varsIgnorePattern": "^_",
+            "args": "after-used",
+            "argsIgnorePattern": "^_"
+        }
+    ],
+
+    "indent": "off",
+    "@typescript-eslint/indent": "error"
+	}
 }
