+++ package.json
@@ -17,6 +17,7 @@
     "antd": "^5.17.4",
     "date-fns": "^3.6.0",
     "next": "14.2.3",
+    "node-fetch": "^3.3.2",
     "react": "18.2.0",
     "react-dom": "^18"
   },
@@ -26,9 +27,11 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@typescript-eslint/eslint-plugin": "^7.11.0",
     "autoprefixer": "^10.4.19",
     "eslint": "^8",
     "eslint-config-next": "14.2.3",
+    "eslint-plugin-unused-imports": "^4.0.0",
     "husky": "^9.0.11",
     "next-sitemap": "^4.2.3",
     "postcss": "^8.4.38",
