+++ src/components/Modal/index.tsx
@@ -1,95 +1,94 @@
-import {SELECT_OPTION} from "@/constant/layout";
 import {getItem, setItem} from "@/utils/localStorage";
-import {Divider, Flex, InputNumber, Select, Modal as _Modal} from "antd";
+import {Divider, Flex, Modal as _Modal} from "antd";
 import {useRouter} from "next/router";
 import {useState} from "react";
 import Content from "./components/Content";
 
 interface Props {
-  isModalOpen: boolean;
-  handleClose: () => void;
+    isModalOpen: boolean;
+    handleClose: () => void;
 }
 
 export interface Info {
-  option: string;
-  wage: number;
+    option: string;
+    wage: number;
 }
 
 export interface Time {
-  time: string;
+    time: string;
 }
 
 export interface Name {
-  name: string;
+    name: string;
 }
 
 const Modal = (props: Props) => {
-  const {isModalOpen, handleClose} = props;
+    const {isModalOpen, handleClose} = props;
 
-  const router = useRouter();
+    const router = useRouter();
 
-  const info = getItem<Info>("info");
+    const info = getItem<Info>("info");
 
-  const [option, setOption] = useState(info?.option);
-  const [wage, setWage] = useState<number | null>(info?.wage || null);
-  const [name, setName] = useState("");
+    const [option, setOption] = useState(info?.option);
+    const [wage, setWage] = useState<number | null>(info?.wage || null);
+    const [name, setName] = useState("");
 
-  const onChangeOption = (value: string) => {
-    setOption(value);
-  };
+    const onChangeOption = (value: string) => {
+        setOption(value);
+    };
 
-  const onChangeWage = (value: number | null) => {
-    setWage(value);
-  };
+    const onChangeWage = (value: number | null) => {
+        setWage(value);
+    };
 
-  const onChangeName = (value: string) => {
-    setName(value);
-  };
+    const onChangeName = (value: string) => {
+        setName(value);
+    };
 
-  const onOk = () => {
-    setItem("info", {option, wage});
-    setItem("time", {time: new Date()});
-    setItem("name", {name});
+    const onOk = () => {
+        setItem("info", {option, wage});
+        setItem("time", {time: new Date()});
+        setItem("name", {name});
 
-    handleClose();
-    router.reload();
-  };
+        handleClose();
+        router.reload();
+    };
 
-  const isDisabled = !option || !wage || !name;
+    const isDisabled = !option || !wage || !name;
 
-  return (
-    <>
-      <_Modal
-        title='하우 머치?'
-        centered
-        okText='저장'
-        cancelText='취소'
-        closable={false}
-        open={isModalOpen}
-        okButtonProps={{
-          disabled: isDisabled,
-        }}
-        onCancel={info && handleClose}
-        onOk={onOk}
-        footer={(_, {OkBtn, CancelBtn}) => (
-          <Flex justify='center' gap={10}>
-            {info ? (
-              <>
-                <CancelBtn />
-                <OkBtn />
-              </>
-            ) : (
-              <OkBtn />
-            )}
-          </Flex>
-        )}
-      >
-        <Divider />
-        <Content option={option} wage={wage} handleChangeOption={onChangeOption} handleChangeWage={onChangeWage} handleChangeName={onChangeName} />
-        <Divider />
-      </_Modal>
-    </>
-  );
+    return (
+        <>
+            <_Modal
+                title='하우 머치?'
+                centered
+                okText='저장'
+                cancelText='취소'
+                closable={false}
+                open={isModalOpen}
+                okButtonProps={{
+                    disabled: isDisabled,
+                }}
+                onCancel={info && handleClose}
+                onOk={onOk}
+                footer={(_, {OkBtn, CancelBtn}) => (
+                    <Flex justify='center' gap={10}>
+                        {info ? (
+                            <>
+                                <CancelBtn />
+                                <OkBtn />
+                            </>
+                        ) : (
+                            <OkBtn />
+                        )}
+                    </Flex>
+                )}
+            >
+                <Divider />
+                <Content option={option} wage={wage} handleChangeOption={onChangeOption} handleChangeWage={onChangeWage} handleChangeName={onChangeName} />
+                <Divider />
+            </_Modal>
+        </>
+    );
 };
 
 export default Modal;
