+++ src/components/layout/Footer.tsx
@@ -2,7 +2,7 @@ import {UpOutlined} from "@ant-design/icons";
 import {FloatButton} from "antd";
 
 const Footer = () => {
-  return <FloatButton className='right-6 bottom-5' icon={<UpOutlined />} />;
+    return <FloatButton className='right-6 bottom-5' icon={<UpOutlined />} />;
 };
 
 export default Footer;
