+++ src/instrumentation.ts
@@ -1,9 +1,9 @@
 export async function register() {
-  if (process.env.NEXT_RUNTIME === 'nodejs') {
-    await import('../sentry.server.config');
-  }
+    if (process.env.NEXT_RUNTIME === 'nodejs') {
+        await import('../sentry.server.config');
+    }
 
-  if (process.env.NEXT_RUNTIME === 'edge') {
-    await import('../sentry.edge.config');
-  }
+    if (process.env.NEXT_RUNTIME === 'edge') {
+        await import('../sentry.edge.config');
+    }
 }
