+++ src/pages/_document.tsx
@@ -2,39 +2,39 @@ import {StyleProvider, createCache, extractStyle} from "@ant-design/cssinjs";
 import Document, {Html, Head, Main, NextScript, DocumentContext} from "next/document";
 
 export default function MyDocument() {
-  return (
-    <Html lang='kr'>
-      <Head />
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
+    return (
+        <Html lang='kr'>
+            <Head />
+            <body>
+                <Main />
+                <NextScript />
+            </body>
+        </Html>
+    );
 }
 
 MyDocument.getInitialProps = async (ctx: DocumentContext) => {
-  const cache = createCache();
-  const originalRenderPage = ctx.renderPage;
-  ctx.renderPage = () =>
-    originalRenderPage({
-      enhanceApp: (App) => (props) =>
-        (
-          <StyleProvider cache={cache}>
-            <App {...props} />
-          </StyleProvider>
-        ),
-    });
+    const cache = createCache();
+    const originalRenderPage = ctx.renderPage;
+    ctx.renderPage = () =>
+        originalRenderPage({
+            enhanceApp: (App) => (props) =>
+                (
+                    <StyleProvider cache={cache}>
+                        <App {...props} />
+                    </StyleProvider>
+                ),
+        });
 
-  const initialProps = await Document.getInitialProps(ctx);
-  const style = extractStyle(cache, true);
-  return {
-    ...initialProps,
-    styles: (
-      <>
-        {initialProps.styles}
-        <style dangerouslySetInnerHTML={{__html: style}} />
-      </>
-    ),
-  };
+    const initialProps = await Document.getInitialProps(ctx);
+    const style = extractStyle(cache, true);
+    return {
+        ...initialProps,
+        styles: (
+            <>
+                {initialProps.styles}
+                <style dangerouslySetInnerHTML={{__html: style}} />
+            </>
+        ),
+    };
 };
