--- src/pages/api/index.page.js
@@ -1,31 +1,0 @@
-import fetch from "node-fetch";
-import {format, subDays} from "date-fns";
-import {wrapApiHandlerWithSentry} from "@sentry/nextjs";
-import https from "https";
-
-async function handler(request, response) {
-  const yesterDay = format(subDays(new Date(), 1), "yyyyMMdd");
-
-  const httpsAgent = new https.Agent({
-    rejectUnauthorized: false,
-  });
-
-  const res = await fetch(`https://www.koreaexim.go.kr/site/program/financial/exchangeJSON?authkey=${process.env.FINANCIAL_AUTH_TOKEN}&data=AP01&searchdate=${yesterDay}`, {
-    method: "GET",
-    headers: {
-      Accept: "application/json",
-      "Content-Type": "application/json",
-    },
-    agent: httpsAgent,
-  });
-  const data = await res.json();
-  try {
-    return response.status(200).json({
-      data,
-    });
-  } catch (e) {
-    return response.status(500).json({error: error});
-  }
-}
-
-export default wrapApiHandlerWithSentry(handler, "/api");
