+++ src/utils/localStorage.ts
@@ -1,14 +1,13 @@
 export const setItem = (key: string, value: object) => {
-  if (typeof window !== 'undefined') {
-  localStorage.setItem(key, JSON.stringify(value))
-  }
-}
+    if (typeof window !== "undefined") {
+        localStorage.setItem(key, JSON.stringify(value));
+    }
+};
 
 export const getItem = <T>(key: string) => {
-  if (typeof window !== 'undefined') {
-  const value = localStorage.getItem(key);
+    if (typeof window !== "undefined") {
+        const value = localStorage.getItem(key);
 
-  
-  return JSON.parse(value!) as T;
-  }
-}
+        return JSON.parse(value!) as T;
+    }
+};
